requirements.txt
numpy==1.26.4
pillow==10.3.0
python-dotenv==1.0.1
